finished test script
