--- pyproj.toml
@@ -1,3 +1,0 @@
-[build-system]
-requires = ["setuptools>=51"]
-build-backend = "setuptools.build_meta"
